net/ppp: Support PPP link layer over serial
Add PPP to Zephyr net stack as data link layer. For now only PPP over
serial is implemented and tested. Also there is only support for IPv4
for now. There is support for PAP authentication as well.

From now it is possible to setup IP connection using cellular
modems. AT commands need to be issued as part of 'connect' callback,
which is supplied by application code.

Signed-off-by: Marcin Niestroj <m.niestroj@grinn-global.com>