net: tcp: Cancel fin_timer in FIN_WAIT2 instead FIN_WAIT1
According to RFC 793 we should wait for FIN in FIN-WAIT-1 and
FIN-WAIT-2 states. Receiving ACK in FIN-WAIT-1 just moves us to
FIN-WAIT-2 state.

Right now TCP connection is never closed if FIN is not received
in FIN-WAIT-2 state. Fix that by keeping fin_timer active in
FIN-WAIT-2 state, but canceling it just after FIN is received.

Fixes: 124c06702726 ("net: tcp: Cancel the fin_timer on FIN message in FIN_WAIT1 state")
Signed-off-by: Marcin Niestroj <m.niestroj@grinn-global.com>